Completed Assignment2
Well, I started out doing mostly what was intended. By trying to
“optimize” the functionality, I seem to have made the second function
not so necessary. The assumption leading to this can be found in the
comments at the head of the code. In short, however, I assumed that
someone would use the inverse, so it may as well be calculated when the
object was created. Thus the second function doesn’t add any
functionality that I can see easily. It does, however, still interact
the way it was supposed to, more or less.